requirements.txt
# DrissionPage现代化自动化框架依赖
# 核心依赖
DrissionPage>=4.0.0
pydantic>=2.0.0
loguru>=0.7.0
PyYAML>=6.0

# 测试框架
pytest>=7.0.0
pytest-html>=3.1.0
pytest-xdist>=3.0.0
pytest-asyncio>=0.21.0
allure-pytest>=2.12.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=9.0.0

# 工具库
requests>=2.28.0
faker>=18.0.0
python-dotenv>=1.0.0
rich>=13.0.0

# 开发工具
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
